added bcrypt

+++ app/controllers/sessions_controller.rb
@@ -12,7 +12,8 @@ class SessionsController < ApplicationController
 
     def create
         if User.exists?(user_params)
-            @user = User.find_by(user_params)
+            @user = User.find_by(username: params[:username])
+            return head(:forbidden) unless @user.authenticate(params[:password])
             session[:user_id] = @user.id
             redirect_to session.delete(:return_to)
         else
@@ -24,12 +25,6 @@ class SessionsController < ApplicationController
         session[:return_to] ||= request.referer
         session.delete :user_id
         redirect_to session.delete(:return_to)
-    end
-
-    private
-
-    def user_params
-        params.permit(:username, :password)
     end
 
 end
